requirements.txt
selenium==4.15.2
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.25
